TranscriptionApp/requirements.txt
numpy<2.0.0
torch>=2.0.0
openai-whisper>=20231117
tqdm>=4.65.0 